Fixed relative paths resolving incorrectly when minifying
Basically there's a bug with the minifier we're using (that's like the recommended .NET core minifier) when it deploys on a Linux machine.

The work around is to only ever reference images from a cshtml file where the path will be resolved correctly. Unfortunately this means that since we want to set a background image in our CSS, we have to do it inline in the layout file.

I don't like it, but it isn't terrible if we limit the CSS not in the CSS file to the bare minimum as I have it right now.